@@ -10,7 +10,12 @@
     "bluebird": "^3.3.4",
     "body-parser": "^1.15.0",
     "bookshelf": "^0.9.4",
+    "connect-redis": "^3.0.2",
+    "cookie-parser": "^1.4.1",
     "express": "^4.13.4",
+    "express-session": "^1.13.0",
+    "express-sessions": "^1.0.5",
+    "flash": "^1.1.0",
     "griddle-react": "^0.4.1",
     "jquery": "^2.2.3",
     "jwt-simple": "^0.5.0",
@@ -18,36 +23,58 @@
     "letsencrypt": "^1.4.3",
     "letsencrypt-express": "^1.1.3",
     "lodash": "^4.11.1",
+    "method-override": "^2.3.5",
     "moment": "^2.12.0",
+    "multer": "^1.1.0",
     "mysql": "^2.10.2",
     "nodemon": "^1.9.1",
+    "passport": "^0.3.2",
+    "passport-github2": "^0.1.10",
     "react": "^0.14.8",
-    "react-bootstrap": "^0.28.5",
+    "react-bootstrap": "^0.29.4",
     "react-dom": "^0.14.8",
+    "react-dropzone": "^3.4.0",
     "react-router": "^2.0.1",
+    "react-select": "^1.0.0-beta13",
+    "redis": "^2.6.0-2",
+    "sharp": "^0.14.1",
     "spdy": "^3.2.3",
     "superagent": "^1.8.3"
   },
   "devDependencies": {
     "babel-core": "^6.7.4",
+    "babel-eslint": "^6.0.3",
     "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-register": "^6.7.2",
     "chai": "^3.5.0",
     "css-loader": "^0.23.1",
+    "eslint": "^2.8.0",
+    "eslint-config-airbnb": "^8.0.0",
+    "eslint-plugin-import": "^1.5.0",
+    "eslint-plugin-jsx-a11y": "^1.0.2",
+    "eslint-plugin-react": "^5.0.1",
+    "file-loader": "^0.8.5",
     "gulp": "^3.9.1",
     "gulp-nodemon": "^2.0.6",
     "gulp-shell": "^0.5.2",
     "http-server": "^0.8.5",
     "mocha": "^2.4.5",
     "mock-local-storage": "^1.0.2",
+    "morgan": "^1.7.0",
+    "session-file-store": "^0.2.0",
+    "style-loader": "^0.13.1",
     "supertest": "^1.2.0",
-    "webpack": "^1.12.14",
-    "webpack-dev-server": "^1.14.1"
+    "webpack": "^1.13.0",
+    "webpack-dev-middleware": "^1.6.1",
+    "webpack-dev-server": "^1.14.1",
+    "webpack-hot-middleware": "^2.10.0"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack && nodemon server/app.js",
+    "test": "mocha -r mock-local-storage test/",
+    "start": "NODE_ENV=development nodemon ./server/app.js --ignore client --history-api-fallback",
     "start:dev": "NODE_ENV=development webpack-dev-server --inline --content-base public/ --history-api-fallback",
     "start:prod": "webpack && node server/app.js",
     "deploy": "NODE_ENV=production webpack -p"
