@@ -1,18 +1,26 @@
 import React from 'react'
 import { Grid } from 'react-bootstrap'
-import Sidebar from './helpers/Sidebar'
+import Sidebar from '../helpers/Sidebar'
+import NavigationBar from '../helpers/NavigationBar.js'
+
 
 
 class Dashboard extends React.Component {
+
+  componentWillMount() {
+
+  }
+  
   render() {
     return(
       <div id="wrapper">
-
+        <NavigationBar isAdminBar={true}/>
         <Sidebar />
-
-        <div id="page-content-wrapper">
-          <div className="container-fluid">
-            {this.props.children}
+        <div id="main">
+          <div id="page-content-wrapper">
+            <div className="container-fluid">
+              {this.props.children}
+            </div>
           </div>
         </div>
 
