@@ -1,10 +1,15 @@
 import React from 'react'
+import ProfileSidebar from './ProfileSidebar'
+import ProfileImage from './ProfileImage'
 import ProfileField from './ProfileField'
 import ProfileSite from './ProfileSite'
+import ProfileGroups from './ProfileGroups'
+import ProfileAdminOptions from './ProfileAdminOptions'
 import ProfileEditButton from './ProfileEditButton'
-// import Image from './Image.js'
+import auth from '../../../util/authHelpers.js'
 import RestHandler from '../../../util/RestHandler'
-import { Button, Row, Col, Image} from 'react-bootstrap';
+import {findDOMNode} from 'react-dom'
+import { InputGroup, ControlLabel, FormControl, Select, Button, Grid, Row, Col, Image} from 'react-bootstrap';
 var _map = require('lodash/map');
 var _find = require('lodash/find');
 var _clone = require('lodash/clone')
@@ -15,20 +20,29 @@ class Profile extends React.Component {
     super (props);
     this.state = {
       profileData: {},
-      editing: 0
+      editing: 0,
     };
 
     //if profile is edited / saved. sites and userInfo into will
     // be converted to an arr of objects -- see saveUserProfile()
-    this.profileEdits = {
-      user: {},
-      sites: {},
-      userInfo:{}
+    this.stagedProfileEdits = {user: {}, sites: {}, userInfo:{}};
+
+  }
+
+  componentWillReceiveProps(nextProps) {
+    //For when the route changes from something like /users/3 to /users/6
+    //When clicking on the sidebar url
+    if (nextProps) {
+      this.state = {
+        profileData: {},
+        editing: 0
+      }
+      this.getUserProfile(nextProps.params.user);
     }
   }
 
-  componentDidMount() {
-    this.getUserProfile();
+  componentWillMount() {
+    this.getUserProfile(this.props.params.user);
   }
 
 
@@ -36,16 +50,18 @@ class Profile extends React.Component {
   // 1. Gets the users profile (but doesn't update the state yet)
   // 2. Run functions to get all the available fields on the site and splice in
   //    the filled out fields to a profileData object
-  // 3. set the state to profileData, which renders most of the page!
-  // 4. Also initialize this.profileEdits.user incase something is edited
-
-  getUserProfile() {
-    var url = '/db/users/user/' + this.props.params.user;
+  // 3. sets the state to profileData, which renders the page!
+  // 4. initializes this.stagedProfileEdits.user to currently viewing user
+  getUserProfile(userId) {
+    var url = '/db/users/user/' + userId;
     RestHandler.Get(url, (err, res) => {
-
+      if (res.body === 'Permission Denied' || !res.body) {
+        window.location.href = '/';
+      }
       this.spliceFilledOutFieldsIntoAvailableFields(res.body, 'userInfo', '/db/fields', (profileData) => {
         this.spliceFilledOutFieldsIntoAvailableFields(profileData, 'sites', '/db/sites', (profileData) => {
-          this.profileEdits.user = res.body.user
+          this.stagedProfileEdits.user = profileData.user;
+          this.stagedProfileEdits.groups = _map(profileData.groups, (key, val) => val).join(',')
           this.setState({
             profileData: profileData,
           });
@@ -54,25 +70,24 @@ class Profile extends React.Component {
     });
   }
 
-
-
   spliceFilledOutFieldsIntoAvailableFields(profileData, objectToUpdate, url, callback) {
     RestHandler.Get(url, (err, res) => {
       var filledOutFields = profileData[objectToUpdate];
       var availableFields = res.body;
-      var newObjectWithAllFields = _map(availableFields,(availableField) => {
-        //sets available field to filled out field if it exists
-        var found = _find(filledOutFields, (field) => field.id === availableField.id);
-        return found ? found : availableField;
+      availableFields.forEach(function(field, i) {
+        if (!filledOutFields[field.id]) {
+          filledOutFields[field.id] = field;
+        }
       });
-      profileData[objectToUpdate] = newObjectWithAllFields
+      profileData[objectToUpdate] = filledOutFields
       callback(profileData)
     });
   }
 
   renderProfileFields() {
     if(this.state.profileData.userInfo) {
-      return this.state.profileData.userInfo.map((detail, index) => {
+      return _map(this.state.profileData.userInfo, (detail, index) => {
+        detail.id = index;
         return (<ProfileField
           fieldDetails={detail}
           editing={this.state.editing}
@@ -81,106 +96,146 @@ class Profile extends React.Component {
       });
     }
   }
+
   renderProfileSites() {
     if(this.state.profileData.sites) {
-      return this.state.profileData.sites.map((site, index) => {
-        return (
-          <li key={index}>
-            <ProfileSite
-            siteDetails={site}
-            editing={this.state.editing}
-            stageProfileEdits = {this.stageProfileEdits.bind(this)} />
-          </li>
-        );
+      return _map(this.state.profileData.sites, (site, index) => {
+        if (site !== undefined) {
+          site.id = index;
+          return (
+            <li key={index}>
+              <ProfileSite
+              siteDetails={site}
+              editing={this.state.editing}
+              stageProfileEdits = {this.stageProfileEdits.bind(this)} />
+            </li>
+          );
+        }
       });
     }
   }
 
+  renderProfileGroups() {
+    if(this.state.profileData.groups) {
+      return (
+      <ProfileGroups
+      selectedGroups={this.state.profileData.groups}
+      editing={this.state.editing}
+      stageProfileEdits = {this.stageProfileEdits.bind(this)} />)
+    }
+  }
+
+  renderProfileAdminOptions() {
+    if (auth.getCookie('ac') === '1' && this.state.editing === 1 && this.state.profileData.user) {
+      return(
+        <ProfileAdminOptions
+          stageProfileEdits = {this.stageProfileEdits.bind(this)}
+          permission = {this.state.profileData.user.permission}
+          public = {this.state.profileData.user.public}
+          userid = {this.state.profileData.user.id}
+          />
+      )
+    }
+  }
 
 
   //In saveUserProfile, the api accepts post requests in array format for
-  //data.userInfo and data.sites. I set them up as an object initially because
+  //data.userInfo and data.sites.
   //it's easier to work worth and then convert it to an array before saving.
   saveUserProfile(callback) {
     var url = '/db/users/user/' + this.props.params.user;
-    var data = this.profileEdits;
+    var data = this.stagedProfileEdits;
     data.userInfo = _map(data.userInfo, function(val){return val});
     data.sites = _map(data.sites, function(val){return val});
     RestHandler.Post(url, data, (err, res) => {
-      console.log(data);
       if (err) {return err;}
       callback(res);
     });
   }
 
   profileEditButtonTapped() {
-    this.state.editing
-      ? this.setState({ editing: 0})
-      : this.setState({ editing: 1})
+    if(this.state.editing) {
+      this.setState({
+        editing: 0,
+      });
+    } else {
+      this.setState({ editing: 1 });
+    }
   }
 
   profileSaveButtonTapped() {
     this.saveUserProfile( () => {
-      this.setState({ editing: 0})
-      this.profileEdits = {
-        user: this.profileEdits.user,
-        sites: {},
-        userInfo:{}
-      }
+      this.stagedProfileEdits.sites = {};
+      this.stagedProfileEdits.userInfo = {};
+      this.setState({ editing: 0 });
+      this.getUserProfile(this.props.params.user);
     });
   }
 
-
-
   // stageProfileEdits is passed into child components where the value can be
-  // edited and saved. It updateds the profileEdits property, which will be
+  // edited and saved. It updates the stagedProfileEdits property, which will be
   // saved when you call saveUserProfile()
-
   stageProfileEdits(callback) {
-    callback(this.profileEdits);
-    // //uncomment to see what's being staged when you edit a field
-    // console.log(this.profileEdits);
+    callback(this.stagedProfileEdits);
+    // console.log("permission: ", this.stagedProfileEdits.user.permission);
+    // console.log("public: ", this.stagedProfileEdits.user.public);
   }
 
   render() {
-    var username, image, group, id = ''
-    if (this.state.profileData.user) {
-      var {username, image, group, id} = this.state.profileData.user
+    var {name, image, id} = (this.state.profileData.user) ? this.state.profileData.user : ''
+    var groups = this.state.profileData.groups;
+    var profileSidebar = groups ? <ProfileSidebar groups={groups} /> : <div></div>;
+    if (this.state.profileData.user && (auth.getCookie('ac') === '1' || this.state.profileData.user.public === 1)) {
+      return (
+        <Row>
+          <Col xs={12} sm={9} xl={10} className='float-right'>
+            <div className='section profile-main'>
+              <Row>
+                <Col xs={12} sm={5} md={4}>
+                  <ProfileImage
+                    src={image}
+                    profileUserId={id}
+                    editing={this.state.editing} />
+                </Col>
+                <Col xs={12} sm={7} md={8}>
+                  <ProfileEditButton
+                    profilesUserId={id}
+                    editing={this.state.editing}
+                    profileEditButtonTapped={this.profileEditButtonTapped.bind(this)}
+                    profileSaveButtonTapped={this.profileSaveButtonTapped.bind(this)} />
+                  <h2>{name}</h2>
+                  {this.renderProfileGroups()}
+
+                  <ul>
+                    {this.renderProfileSites()}
+                  </ul>
+                  {this.renderProfileAdminOptions()}
+                </Col>
+              </Row>
+            </div>
+            <div className='section profile-bio'>
+              {this.renderProfileFields()}
+              <ProfileEditButton
+                profilesUserId = {id}
+                editing = {this.state.editing}
+                hideEditButton = {true}
+                profileEditButtonTapped = {this.profileEditButtonTapped.bind(this)}
+                profileSaveButtonTapped = {this.profileSaveButtonTapped.bind(this)} />
+            </div>
+          </Col>
+          <Col xs={12} sm={3} xl={2}>
+            {profileSidebar}
+          </Col>
+        </Row>
+
+      );
+    } else {
+      return (
+        <div>
+        {/*Sorry, this user's profile is not publicly visible*/}
+        </div>
+      );
     }
-
-    return (
-      <div>
-        <div className='section profile-main'>
-            <Row>
-              <Col xs={12} sm={5} md={4}>
-                <Image src={image} responsive />
-              </Col>
-              <Col xs={12} sm={7} md={8}>
-                <ProfileEditButton
-                  profilesUserId = {id}
-                  editing = {this.state.editing}
-                  profileEditButtonTapped = {this.profileEditButtonTapped.bind(this)}
-                  profileSaveButtonTapped = {this.profileSaveButtonTapped.bind(this)} />
-                <h2>{username}</h2>
-                <h3>Groups: <a href="#">{group}</a></h3>
-                <ul>
-                  {this.renderProfileSites()}
-                </ul>
-              </Col>
-            </Row>
-        </div>
-        <div className = 'section profile-bio'>
-          {this.renderProfileFields()}
-          <ProfileEditButton
-            profilesUserId = {id}
-            editing = {this.state.editing}
-            hideEditButton = {true}
-            profileEditButtonTapped = {this.profileEditButtonTapped.bind(this)}
-            profileSaveButtonTapped = {this.profileSaveButtonTapped.bind(this)} />
-        </div>
-      </div>
-
-    );
   }
 }
 
