@@ -1,37 +1,43 @@
-import React from 'react'
+import React from 'react';
 import { Button, ButtonToolbar } from 'react-bootstrap';
-import auth from '../../../util/authHelpers.js'
-
+import auth from '../../../util/authHelpers.js';
+import RestHandler from '../../../util/RestHandler.js';
 
 class ProfileEditButton extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      loggedInUserId: -1,
-      permission: 0
-    }
+      loggedInUserId: -666,
+      permission: auth.getCookie('ac'),
+    };
   }
-
-  componentDidMount() {
-    this.getLoggedInUserData();
+  
+  componentWillMount() {
+    RestHandler.Get('/db/users/name', (err, res) => {
+      if (err) {
+        console.log(err);
+      } else {
+        this.setState({ loggedInUserId: res.body.id });
+      }
+    });
   }
 
   getLoggedInUserData() {
     //gets the authentication token from util/authHelpers.js
     //and retrieves permission and user id
-    auth.parseJwtAsync((parsedToken) => {
-      if(parsedToken) {
-        this.setState({
-          loggedInUserId: parsedToken.iss,
-          permission: parsedToken.perm
-        });
-      }
-    });
+    // auth.parseJwtAsync((parsedToken) => {
+    //   if(parsedToken) {
+    //     this.setState({
+    //       loggedInUserId: parsedToken.iss,
+    //       permission: parsedToken.perm
+    //     });
+    //   }
+    // });
   }
 
 
   generateButtons() {
-    if(this.state.permission === 1 || this.props.profilesUserId === this.state.loggedInUserId) {
+    if(this.state.permission === '1' || this.props.profilesUserId === this.state.loggedInUserId) {
       if(this.props.editing) {
         return (
           <ButtonToolbar>
