@@ -6,26 +6,42 @@ var Get = function(url, callback) {
     .get(url)
     .end((err, res) => {
       err
-        ? console.error('Err in util/restHander ', err)
+        ? callback(err, res)
         : callback(err, res);
     });
 }
 
 var Post = function (url, data, callback) {
-  if(localStorage.getItem('jwtAlum')){
-    data.token = JSON.parse(localStorage.getItem('jwtAlum')).token;
-  }
+  // if(localStorage.getItem('jwtAlum')){
+  //   data.token = JSON.parse(localStorage.getItem('jwtAlum')).token;
+  // }
   request
     .post(url)
     .send(data)
     .end((err, res) => {
-      err
-        ? console.error('Err in util/restHander ', err)
-        : callback(err, res);
+      if(err) {
+        console.error('Err in util/restHander ', err);
+      }
+      callback(err, res);
+  });
+}
+
+var Delete = function (url, callback) {
+  // if(localStorage.getItem('jwtAlum')){
+  //   data.token = JSON.parse(localStorage.getItem('jwtAlum')).token;
+  // }
+  request
+    .delete(url)
+    .end((err, res) => {
+      if(err) {
+        console.error('Err in util/restHander ', err);
+      }
+      callback(err, res);
   });
 }
 
 module.exports = {
   Get: Get,
-  Post: Post
+  Post: Post,
+  Delete: Delete
 };
