@@ -2,6 +2,7 @@ var bodyParser = require('body-parser');
 var util = require('./utility.js');
 var Promise = require('bluebird');
 var User = require('../models/user.js');
+var bcrypt = require('bcrypt');
 
 module.exports = {
   getAll: function(attrs, Model) {
@@ -31,28 +32,4 @@ module.exports = {
     });
   },
 
-  checkLogin: function(req, res) {
-
-    User.where({email: req.body.email}).fetch().then(function(user){
-      if(user === null) {
-        res.send(401, 'No user with that email');
-      } else {
-        // bcrypt.compare(password, user[0].password, function(err, isMatch){
-        //   if(match) {
-        //     res.send(200, util.generateToken(user.id, user));
-        //   } else {
-        //     // res.send(401, 'Invalid Password');
-        //     res.send(200, util.generateToken(user.id, user));
-
-        //   }
-        // });
-        if(req.body.password === user.attributes.password) {
-          res.send(200, JSON.stringify({token: util.generateToken(user.attributes.id, 
-            user.attributes.email, user.attributes.permission, user.attributes.username)}));
-        } else {
-          res.send(403, 'Invalid Password');
-        }
-      }
-    });
-  }
 };
