@@ -3,7 +3,10 @@ var db = require('../dbConfig');
 var Group = db.Model.extend({
   tableName: 'Groups',
   users: function() {
-    return this.hasMany('User');
+    return this.belongsToMany('User');
+  },
+  visibleGroups: function() {
+    return this.hasMany('VisibleGroup');
   }
 });
 
