@@ -3,7 +3,8 @@ var db = require('../dbConfig');
 var User = db.Model.extend({
   tableName: 'Users',
   groups: function() {
-    return this.belongsTo('Group');
+    // return this.belongsTo('Group'); // changed!!
+    return this.belongsToMany('Group');
   },
   userSites: function() {
     return this.hasMany('UserSite');
