@@ -3,11 +3,11 @@ var groupController = require('../controllers/groupController');
 var util = require('../lib/utility.js');
 
 groupRouter.route('/')
-  .get(groupController.fetchGroups)
+  .get(util.isLoggedIn, groupController.fetchGroups2)
   .post(util.isAdmin, groupController.createGroup);
 
 groupRouter.route('/group/:id')
-  .get(groupController.fetchGroupInfo)
+  .get(util.isLoggedIn, groupController.fetchGroupInfo2)
   .post(util.isAdmin, groupController.modifyGroup)
   .delete(util.isAdmin, groupController.deleteGroup);
 
