@@ -1,22 +1,35 @@
+const webpack = require('webpack');
+const path = require('path');
+
 module.exports = {
-
-  entry: './client/index.js',
+  devtool: 'cheap-module-eval-source-map',
+  entry: [
+    'webpack-hot-middleware/client',
+    './client/index.js'],
   output: {
-    path: process.env.NODE_ENV === 'production' ? './client/dist' : './client/build',
-    filename: '/bundle.js'
+    path: path.resolve('./build'),
+    filename: 'bundle.js',
+    publicPath: '/',
   },
-
+  plugins: [
+    new webpack.optimize.OccurrenceOrderPlugin(),
+    new webpack.HotModuleReplacementPlugin(),
+    new webpack.NoErrorsPlugin(),
+  ],
   module: {
     loaders: [
-      { test: /\.js$/,
-        exclude: /node_modules/,
+      {
+        test: /\.js$/,
         loader: 'babel-loader',
-        query:
-        {
-          presets:['es2015', 'react']
-        }
-      }
-
+        exclude: /node_modules/,
+        query: {
+          presets: ['react', 'es2015', 'react-hmre'],
+          cacheDirectory: true
+        },
+      },
+      { test: /\.css$/, exclude: /\.useable\.css$/, loader: "style!css" },
+      { test: /\.useable\.css$/, loader: "style/useable!css" }
     ]
   }
-}
+
+};
